+++ spring-boot-extension-autoconfigure/src/main/java/com/livk/autoconfigure/mybatis/EnableSqlMonitor.java
@@ -16,7 +16,7 @@
 
 package com.livk.autoconfigure.mybatis;
 
-import com.livk.commons.spring.AutoImport;
+import com.livk.commons.annotation.AutoImport;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
