+++ spring-extension-commons/src/test/java/com/livk/commons/annotation/AutoImportSelectorTests.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.livk.commons.spring;
+package com.livk.commons.annotation;
 
 import com.livk.auto.service.annotation.SpringAutoService;
 import org.junit.jupiter.api.Test;
@@ -30,34 +30,30 @@ import static org.assertj.core.api.Assertions.assertThat;
 /**
  * @author livk
  */
-class SpringAbstractImportSelectorTests {
+class AutoImportSelectorTests {
+
+	final AutoImportSelector importSelector = new AutoImportSelector();
 
 	@Test
-	void testFindAnnotation() {
-		MyAnnotationImportSelector selector = new MyAnnotationImportSelector();
-		assertThat(selector.getAnnotationClass()).isEqualTo(MyAnnotation.class);
-
-		String[] imports = selector.selectImports(AnnotationMetadata.introspect(Config.class));
-		String[] result = new String[] { MyAnnotationConfig.class.getName() };
+	void test() {
+		String[] imports = importSelector.selectImports(AnnotationMetadata.introspect(Config.class));
+		String[] result = new String[] { ImportConfig.class.getName() };
 		assertThat(imports).containsExactly(result);
 	}
 
+	@AutoImport
 	@Target(ElementType.TYPE)
 	@Retention(RetentionPolicy.RUNTIME)
-	@interface MyAnnotation {
+	@interface AutoServiceImport {
 
 	}
 
-	static class MyAnnotationImportSelector extends SpringAbstractImportSelector<MyAnnotation> {
+	@SpringAutoService(AutoServiceImport.class)
+	static class ImportConfig {
 
 	}
 
-	@SpringAutoService(MyAnnotation.class)
-	static class MyAnnotationConfig {
-
-	}
-
-	@MyAnnotation
+	@AutoServiceImport
 	static class Config {
 
 	}
