+++ spring-extension-context/src/main/java/com/livk/context/disruptor/DisruptorScanRegistrar.java
@@ -16,7 +16,7 @@
 
 package com.livk.context.disruptor;
 
-import com.livk.commons.spring.AnnotationBasePackageSupport;
+import com.livk.commons.annotation.AnnotationBasePackageSupport;
 import com.livk.commons.util.ObjectUtils;
 import com.livk.context.disruptor.exception.DisruptorRegistrarException;
 import org.springframework.beans.factory.support.BeanDefinitionRegistry;
